refactor: 바인딩 바텀 시트 프래그먼트 구현 및 인원 수 프래그먼트 불필요한 로직 정리
