+++ android/app/src/main/java/com/created/team201/presentation/createStudy/CreateStudyActivity.kt
@@ -6,6 +6,7 @@ import androidx.activity.viewModels
 import com.created.team201.R
 import com.created.team201.databinding.ActivityCreateStudyBinding
 import com.created.team201.presentation.common.BindingActivity
+import com.created.team201.presentation.createStudy.bottomSheet.PeopleCountBottomSheetFragment
 
 class CreateStudyActivity :
     BindingActivity<ActivityCreateStudyBinding>(R.layout.activity_create_study) {
@@ -33,7 +34,24 @@ class CreateStudyActivity :
 
     fun onIconTextButtonClick(tag: String) {
         supportFragmentManager.findFragmentByTag(tag)?.let { return }
-        CreateStudyBottomSheetFragment().show(supportFragmentManager, tag)
+
+        when (tag) {
+            getString(R.string.createStudy_tag_people_count) -> {
+                PeopleCountBottomSheetFragment().show(
+                    supportFragmentManager,
+                    tag,
+                )
+            }
+
+            getString(R.string.createStudy_tag_start_date) -> {
+            }
+
+            getString(R.string.createStudy_tag_period) -> {
+            }
+
+            getString(R.string.createStudy_tag_cycle) -> {
+            }
+        }
     }
 
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
