--- android/app/src/main/res/layout/fragment_create_study_bottom_sheet.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:bind="http://schemas.android.com/apk/res-auto">
-
-    <data>
-
-        <variable
-            name="viewModel"
-            type="com.created.team201.presentation.createStudy.CreateStudyViewModel" />
-
-        <variable
-            name="changeListener"
-            type="com.created.team201.presentation.createStudy.custom.PickerChangeListener" />
-    </data>
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <include
-            android:id="@+id/layout_create_study_people_count"
-            layout="@layout/layout_create_study_people_count"
-            bind:changeListener="@{changeListener}"
-            bind:maxValue="@{8}"
-            bind:minValue="@{2}"
-            bind:viewModel="@{viewModel}" />
-
-    </LinearLayout>
-</layout>
